LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/anonymous-iframe/requestStorageAccess.tentative.https.window-expected.txt


FAIL Same-origin credentialless iframe can't request storage access assert_equals: expected "failed" but got "success"
FAIL Cross-origin credentialless iframe can't request storage access assert_equals: expected "failed" but got "success"
